requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
google-generativeai==0.5.0
sentence-transformers==2.2.2
chromadb==0.4.17
pandas==2.1.1
python-dotenv==1.0.0
gdown==4.7.1  
